mlops/requirements.txt
# ============================================================================
# requirements.txt
# ============================================================================
# AirAware MLOps Requirements

# Core ML Libraries
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
optuna==3.5.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
pymongo==4.6.1

# API & Serving
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# MLOps Tools
mlflow==2.9.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
joblib==1.3.2
scipy==1.11.4

# MQTT (for integration)
paho-mqtt==1.6.1
